requirements.txt
accelerate>=0.15.0
diffusers>=0.17.1
safetensors
einops>=0.6.0
hydra-core
loguru
matplotlib
numpy
omegaconf
opencv_python
Pillow>=9,<10
pyarrow
PyYAML
scikit_learn
tqdm
transformers>=4.25.1
xformers==0.0.22
pytorch_lightning
